Move exceptions to common.
