+++ podium-common/src/main/java/nl/thehyve/podium/common/exceptions/ActionNotAllowedInStatus.java
@@ -5,7 +5,7 @@
  * See the file LICENSE in the root of this repository.
  */
 
-package nl.thehyve.podium.exceptions;
+package nl.thehyve.podium.common.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
