+++ podium-gateway/src/main/java/nl/thehyve/podium/service/RequestReviewProcessService.java
@@ -11,8 +11,8 @@ import nl.thehyve.podium.common.request.DecisionOutcome;
 import nl.thehyve.podium.common.request.RequestReviewStatus;
 import nl.thehyve.podium.common.security.AuthenticatedUser;
 import nl.thehyve.podium.domain.RequestReviewProcess;
-import nl.thehyve.podium.exceptions.ActionNotAllowedInStatus;
-import nl.thehyve.podium.exceptions.ResourceNotFound;
+import nl.thehyve.podium.common.exceptions.ActionNotAllowedInStatus;
+import nl.thehyve.podium.common.exceptions.ResourceNotFound;
 import nl.thehyve.podium.repository.RequestReviewProcessRepository;
 import nl.thehyve.podium.repository.search.RequestReviewProcessSearchRepository;
 import org.flowable.engine.HistoryService;
